+++ Task2/2.1/users.ts
@@ -1,0 +1,11 @@
+type User = {
+    id: string;
+    login: string;
+    password: string;
+    age: number;
+    isDeleted: boolean;
+};
+
+const users: Map<string, User> = new Map();
+
+export default users;
